requirements.txt
pyyaml==6.0.2
duckdb==1.1.3
google-generativeai==0.8.3
pandas==2.2.3
requests==2.32.3